requirements.txt
pbr
pyyaml
flask
